Changed an int to a double so that the zoom value is not being truncated. This fixes the issue with saving an image with a zoom level of less than 1.

+++ carta/cpp/core/Data/Image/Controller.cpp
@@ -1017,7 +1017,7 @@ void Controller::saveState() {
 
 
 QString Controller::saveImage( const QString& fileName ){
-    int zoomLevel = getZoomLevel();
+    double zoomLevel = getZoomLevel();
     return saveImage( fileName, zoomLevel );
 }
